keep consistency with error message, add assertHeader*, assertRedirect*, move tests controller files
